// src/app/type/type.ts
export interface Product {
    id: number;
    name: string;
    description: string;
    availability: string;
    depart: string;
    oldPrice: string;
    price: string;
    image: string;
    images: string[];
    colors: string[];
    material: string;
    sizeAvailability: string[];
    rating: number;
    category: string;
  }
  